ロジック処理の形式を変更
